+++ src/no_std/pedersen.rs
@@ -12,7 +12,6 @@ use elliptic_curve::{
 };
 use rand_chacha::ChaChaRng;
 use rand_core::{CryptoRng, RngCore, SeedableRng};
-use zeroize::Zeroize;
 
 /// Result from calling Pedersen::split_secret
 #[derive(Clone, Debug)]
@@ -63,7 +62,7 @@ impl<const T: usize, const N: usize> Pedersen<T, N> {
         rng: &mut R,
     ) -> Result<PedersenResult<F, G, S, T, N>, Error>
     where
-        F: PrimeField + Zeroize,
+        F: PrimeField,
         G: Group + GroupEncoding + Default + ScalarMul<F>,
         R: RngCore + CryptoRng,
     {
@@ -74,9 +73,8 @@ impl<const T: usize, const N: usize> Pedersen<T, N> {
         let mut crng = ChaChaRng::from_seed(seed);
 
         let g = share_generator.unwrap_or_else(G::generator);
-        let mut t = F::random(&mut crng);
+        let t = F::random(&mut crng);
         let h = blind_factor_generator.unwrap_or_else(|| G::generator() * t);
-        t.zeroize();
 
         let blinding = blinding.unwrap_or_else(|| F::random(&mut crng));
         let (secret_shares, secret_polynomial) =