+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this crate will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v2.2.0 - 2022-12
+## v2.3.0 - 2022-12-16
+- Relax requirement for implementing zeroize
+
+
+## v2.2.0 - 2022-12-15
 - Update dependencies
 - Error implements Display and Error traits
 - Share can be converted to PrimeField and GroupEncoding